% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{basicStatFunct.gmat}
\alias{basicStatFunct.gmat}
\title{basicStatFunct.gmat}
\usage{
basicStatFunct.gmat(sel_dat)
}
\arguments{
\item{sel_data}{selected user data}
}
\value{
list of objects
}
\description{
Return basic statistics for selected data
}
