% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{checkDS.gmat}
\alias{checkDS.gmat}
\title{checkDS.gmat}
\usage{
checkDS.gmat(df, dfName)
}
\arguments{
\item{df}{data frame}

\item{dfName}{data frame name}
}
\value{
dataframe
}
\description{
Return object type for a selected object
}
