% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{colClasses.gmat}
\alias{colClasses.gmat}
\title{fnObjectsList.gmat}
\usage{
colClasses.gmat(df)
}
\arguments{
\item{df}{data frame}
}
\value{
dataframe
}
\description{
Return columns and class for a data source
}
