% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{colDS.gmat}
\alias{colDS.gmat}
\title{colDS.gmat}
\usage{
colDS.gmat(df, dfName)
}
\arguments{
\item{df}{data frame}

\item{dfName}{data frame name}
}
\value{
dataframe
}
\description{
Return column information for dataframe
}
