% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{conInfoFunct2.gmat}
\alias{conInfoFunct2.gmat}
\title{conInfoFunct2.gmat}
\usage{
conInfoFunct2.gmat(sel_dat, type = "")
}
\arguments{
\item{type}{data type}

\item{sel_data}{selected user data}
}
\value{
variable name
}
\description{
Returns obs missing information
}
