% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{conObject.gmat}
\alias{conObject.gmat}
\title{conObject.gmat function}
\usage{
conObject.gmat(conList)
}
\arguments{
\item{conList}{connection to the database server}
}
\value{
list of db connections
}
\description{
Get list of available db in r environment
}
