% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{convRowName.gmat}
\alias{convRowName.gmat}
\title{convRowName.gmat}
\usage{
convRowName.gmat(rowname)
}
\arguments{
\item{rowname}{row name to convert}
}
\value{
array
}
\description{
convert rownames to numeric is valid (Used in beams View Object)
}
