% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{dsClasses.gmat}
\alias{dsClasses.gmat}
\title{dsClasses.gmat}
\usage{
dsClasses.gmat(df, dfName)
}
\arguments{
\item{df}{data frame}

\item{dfName}{data frame name}
}
\value{
dataframe
}
\description{
Return class and other information for a data source
}
