% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{extract_help.gmat}
\alias{extract_help.gmat}
\title{extract_help.gmat}
\usage{
extract_help.gmat(
  pkg,
  fn = NULL,
  to = c("txt", "html", "latex", "ex"),
  lang = "R"
)
}
\arguments{
\item{pkg}{package name}

\item{fn}{function}

\item{to}{return format}

\item{lang}{language default R}
}
\value{
text, html, latex, ex
}
\description{
Extract help page
}
