% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{extract_helpSearch.gmat}
\alias{extract_helpSearch.gmat}
\title{extract_helpSearch.gmat}
\usage{
extract_helpSearch.gmat(keyWord)
}
\arguments{
\item{keyWord}{keyWord to be used in search}
}
\value{
html
}
\description{
Extract help search results into a variable for display
}
