% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{extract_langhelp.gmat}
\alias{extract_langhelp.gmat}
\title{extract_langhelp.gmat}
\usage{
extract_langhelp.gmat(pkg, href, lang = "R")
}
\arguments{
\item{pkg}{package name}

\item{href}{package location}

\item{lang}{language default R}
}
\value{
html
}
\description{
Extract language help into a variable for display
}
