% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{extract_libhelp.gmat}
\alias{extract_libhelp.gmat}
\title{extract_libhelp.gmat}
\usage{
extract_libhelp.gmat(pkg, lang = "R")
}
\arguments{
\item{pkg}{package name}

\item{lang}{language default R}
}
\value{
html
}
\description{
Extract library help into a variable for display
}
