% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{extract_libnews.gmat}
\alias{extract_libnews.gmat}
\title{extract_libnews.gmat}
\usage{
extract_libnews.gmat(packName)
}
\arguments{
\item{packName}{package name}
}
\value{
text
}
\description{
Extract library news into a variable for display
}
