% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{extract_libpage.gmat}
\alias{extract_libpage.gmat}
\title{extract_searchpage.gmat}
\usage{
extract_libpage.gmat(pkg, href, lang = "R")
}
\arguments{
\item{pkg}{package name}

\item{href}{package location}

\item{lang}{language default R}
}
\value{
text
}
\description{
Extract language search page help into a variable for display
}
