% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{InMemList.gmat}
\alias{InMemList.gmat}
\title{InMemList.gmat}
\usage{
InMemList.gmat(exclude = NULL)
}
\arguments{
\item{exclude}{list of objects to exclude}
}
\value{
list of objects
}
\description{
Get list of in memory objects
}
