% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{InMemObjects.gmat}
\alias{InMemObjects.gmat}
\title{InMemObjects.gmat}
\usage{
InMemObjects.gmat()
}
\value{
list of objects
}
\description{
Get list of in memory objects
}
\examples{
test2<-jsonlite::toJSON(InMemObjects.gmat());
}
