% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{is.finite.data.frame}
\alias{is.finite.data.frame}
\title{is.finite.data.frame function}
\usage{
\method{is.finite}{data.frame}(obj)
}
\arguments{
\item{obj}{Object to test}
}
\value{
true or false
}
\description{
Add is infinite method to data frame object to determine if object is finite.
}
