% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{is.infinite.data.frame}
\alias{is.infinite.data.frame}
\title{is.infinite.data.frame function}
\usage{
\method{is.infinite}{data.frame}(obj)
}
\arguments{
\item{obj}{Object to test}
}
\value{
true or false
}
\description{
Add is infinite method to data frame object to determine if object is infinite.
}
