% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{is.infinite.data.table}
\alias{is.infinite.data.table}
\title{is.infinite.data.table}
\usage{
\method{is.infinite}{data.table}(obj)
}
\arguments{
\item{selected}{object}
}
\value{
true or false
}
\description{
Add is.infinite method to data table object
}
