% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{packsCode.gmat}
\alias{packsCode.gmat}
\title{packsCode.gmat}
\usage{
packsCode.gmat(
  codeTxt,
  comment = TRUE,
  blank = FALSE,
  arrow = TRUE,
  envInfo = FALSE
)
}
\arguments{
\item{codeTxt}{code text to be evaluate}

\item{comment}{keep comments embedded in codeTxt}

\item{blank}{keep blank rows}

\item{arrow}{replace = with <-}

\item{envInfo}{include R environment information}
}
\value{
list
}
\description{
extract and return formated code, packages and enviroment info
}
