% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{PlotFunct.gmat}
\alias{PlotFunct.gmat}
\title{PlotFunct.gmat}
\usage{
PlotFunct.gmat(plotDat, typePlot)
}
\arguments{
\item{plotDat}{data object to be plotted}

\item{typeplot}{name of plot type}
}
\value{
svg
}
\description{
Returns svg object containing the plot type requested
which is required to display graph in beams
}
