% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{toolCode.gmat}
\alias{toolCode.gmat}
\title{toolCode.gmat}
\usage{
toolCode.gmat(packTarg, targVal)
}
\arguments{
\item{packTarg}{package name}

\item{targVal}{function name}
}
\value{
array
}
\description{
return beam canvas object for selected function in a give package
}
