% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{trycatchExec.gmat}
\alias{trycatchExec.gmat}
\title{trycatchReturn.gmat function}
\usage{
trycatchExec.gmat(expr)
}
\arguments{
\item{expr}{}
}
\value{
list of messages, warnings and errors
}
\description{
trycatchExec.gmat is used by beams to capture messages, warnings and
errors when submitting user defined fuctions. Notice that
semi-colons are embedded into the string after each function.
}
\examples{
trycatchExec.gmat({"+rmEqStr+"});
info = return_ERR[2:4];
 rm(list = ls(pattern = '_ERR'));
}
