% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{WrapPlot.gmat}
\alias{WrapPlot.gmat}
\title{WrapPlot.gmat}
\usage{
WrapPlot.gmat(plotfunction)
}
\arguments{
\item{plotfunction}{function that is to be wrapped}
}
\value{
svg
}
\description{
Returns svg object containing the plot type requested
which is required to dsiplay graph in beams
}
