% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamsUtil.R
\name{trycatchReturn.gmat}
\alias{trycatchReturn.gmat}
\title{trycatchReturn.gmat function}
\usage{
trycatchReturn.gmat(expr)
}
\arguments{
\item{expr}{expression provided by user to be executed}
}
\value{
list of messages, warnings and errors
}
\description{
trycatchReturn.gmat is used by beams to capture messages, warnings and
errors when submitting return user defined fuctions. Notice that
semi-colons are embedded into the string after each function.
}
